Python spells "abstract method" as "raise NotImplementedError...".  This patch more-or-less enforces that dictum.  If a "LOAD_GLOBAL NotImplementedError" is seen before the raise statement that triggers a return, assume "raise NotImplementedError" was executed.
This can obviously be fooled:

    def f(self):
	x = NotImplementedError
	raise SystemError

and note that arbitrarily complex expressions can be involved in the raise
statement:

    def f(self):
	raise NotImplementedError, "*"*14 + " msg " + "*"*14

so the distance between LOAD_GLOBAL "NotImplementedError" and RAISE_VARARGS
can vary.  Still, this seems like a better approximation to the truth than
assuming a priori that all raise statements imply abstract methods.